requirements.txt
accelerate==1.6.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
arviz==0.20.0
astropy ## version change
astropy-iers-data==0.2025.6.2.0.38.23
asttokens==3.0.0
attrs==25.1.0
autograd==1.8.0
beautifulsoup4==4.13.3
bm3d==4.0.3
bm4d==4.2.5
camb==1.6.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
corner==2.2.3
coverage==7.8.2
coveralls==4.0.1
cycler==0.12.1
Cython==3.1.1
debugpy==1.8.12
decorator==5.2.0
docopt==0.6.2
ducc0==0.38.0
einops==0.8.1
ema-pytorch==0.7.7
executing==2.2.0
filelock==3.17.0
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
gdown==5.2.0
h5netcdf==1.5.0
h5py==3.13.0
healpy==1.18.1
huggingface-hub==0.30.2
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.32.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.0.post0
lightning-utilities==0.12.0
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pixell==0.28.4
platformdirs==4.3.6
pluggy==1.6.0
prompt_toolkit==3.0.50
propcache==0.3.0
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyerfa==2.0.1.5
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
pytest==8.4.0
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
setuptools==75.8.0
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2025.2.18
torch==2.6.0
torchmetrics==1.6.1
torchsummary==1.5.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
wcwidth==0.2.13
xarray==2025.1.2
xarray-einstats==0.8.0
yarl==1.18.3
